requirements.txt
# Core Dependencies
numpy==1.24.0
pandas==2.0.3
requests==2.31.0
scipy==1.11.3

# Logging & Monitoring
loguru==0.7.0

# Trading & Market Data
yfinance==0.2.30
ccxt==4.0.82  # Crypto Exchange APIs (Binance, Coinbase, etc.)
ib_insync==0.9.84  # Interactive Brokers API

# Machine Learning & AI
scikit-learn==1.3.1
xgboost==1.7.6
tensorflow==2.14.0  # Deep Learning for AI-based strategy
torch==2.1.0  # PyTorch for advanced ML models

# Natural Language Processing (For News Sentiment Analysis)
nltk==3.8.1
transformers==4.34.0
sentence-transformers==2.2.2  # Advanced NLP-based sentiment analysis

# Visualization & Dashboard
matplotlib==3.8.0
tkinterweb==3.6.4  # Advanced UI components for Tkinter
dash==2.14.0  # Web-based monitoring dashboard

# Automation & WebSockets
websockets==11.0.3  # Real-time order execution
apscheduler==3.10.4  # Task scheduling & automated trade reports

# Cloud & Security
python-dotenv==1.0.0  # Secure API key handling
boto3==1.34.0  # AWS integration for cloud storage & monitoring
paramiko==3.4.0  # Secure SSH for remote bot management
cryptography==41.0.7  # Secure data encryption

# PDF & Report Generation
fpdf==1.7.2
reportlab==4.0.8  # Advanced PDF reporting

# Testing & Debugging
pytest==7.4.3
mock==5.1.0
unittest2==1.1.0
